Separate declarations not needed by plugins into cmd.h
