+++ src/daemon/collectd.h
@@ -266,16 +266,4 @@
 
 #include "globals.h"
 
-struct cmdline_config {
-  bool test_config;
-  bool test_readall;
-  bool create_basedir;
-  const char *configfile;
-  bool daemonize;
-};
-
-void stop_collectd(void);
-struct cmdline_config init_config(int argc, char **argv);
-int run_loop(bool test_readall);
-
 #endif /* COLLECTD_H */
